New Resource: azurerm_hdinsight_cluster_pool
